requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing
rembg>=2.0.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Deep Learning (RMBG-2.0)
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
kornia==0.7.1

# Async & Task Queue
celery==5.3.4
redis==5.0.1
aiofiles==23.2.1

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
google-cloud-storage==2.10.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
